Original Video source:
The characters used in this video are the property of Disney. This project is intended solely for educational purposes as part of my game development learning experience. No copyright infringement is intended, and the game is not for commercial use. All rights to the characters and related intellectual property remain with Disney.